# TransportTicketApp/src/main/java/com/github/alekseypetkun/TransportTicketApp/dto/ResponseWrapperTickets.java
package com.github.alekseypetkun.TransportTicketApp.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

/**
 * ДТО всех билетов
 */
@Data
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class ResponseWrapperTickets {

    @Schema(description = "общее количество билетов")
    private Long count;

    @Schema(description = "все найденные билеты")
    private List<TicketDto> results;
}